+++ OverworldChests/LastOverWorldChestSpawn.cs
@@ -1,7 +1,0 @@
-﻿namespace OverworldChests
-{
-	internal class LastOverWorldChestSpawn
-	{
-		public int lastOverworldChestSpawn;
-	}
-}