+++ OverworldChests/release-notes.md
@@ -2,6 +2,11 @@
 
 # Release notes
 
+## 0.5.4-unofficial.2-mouahrara
+Released on 11 April, 2024, for SMAPI 4.0.0 or later.
+* Added console commands: **oc_remove**, **oc_spawn**, and **oc_respawn** to remove, spawn, or respawn overworld chests ✨
+* Overworld chests now always respawn after loading a save 🔧
+
 ## 0.5.4-unofficial.1-mouahrara
 Released on 7 April, 2024, for SMAPI 4.0.0 or later.
 * Migrated to 1.6.0 and project cleanup 🚀